// src/app/shared/generated/model/messageType.ts
/**
 * onecx-chat-ui-bff
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export enum MessageType {
    Human = 'HUMAN',
    Assistant = 'ASSISTANT',
    System = 'SYSTEM'
}